app/requirements.txt
fastapi
sqlalchemy
httpx
python-dotenv
redis
aiosqlite
email-validator
asyncpg
